+++ app/Http/Controllers/MedicineController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Models\Category;
 use App\Models\Expirationdate;
 use App\Models\Medicine;
+use App\Models\User;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Validator;
 
@@ -17,22 +18,36 @@ class MedicineController extends Controller
     }
 
     public function create(Request $request){
+
         $request->validate([
             'scientific_name'=>'required|string|max:25',
             'category_id'=>'required',
-            'trade_name'=>'required|string|max:25',
+            'trade_name'=>'required|string|max:25|unique:medicines',
+            'image'=>'required|image',
             'company'=>'required|string|max:25',
             'price'=>'required|max:25']);
 
-        $medicine=Medicine::create(
-            $request->only('scientific_name','category_id','trade_name','company','price')
-        );
+        $imagePath = $request->file('image')->store('public/images');
 
-        if ($medicine) {
-            return $this->apiResponse($medicine, 'the medicine inserted',201);
-        }
 
-        return $this->apiResponse(null, 'the medicine didn\'t created',404);
+        if(auth()->user()->role){
+            $medicine = Medicine::create([
+                'scientific_name' => $request->input('scientific_name'),
+                'category_id' => $request->input('category_id'),
+                'trade_name' => $request->input('trade_name'),
+                'company' => $request->input('company'),
+                'price' => $request->input('price'),
+                'image' => $imagePath,
+            ]);
+
+
+            if ($medicine) {
+                    return $this->apiResponse($medicine, 'the medicine inserted',201);
+                }
+
+            }
+
+        return $this->apiResponse(null, 'the medicine didn\'t created you arent admin',404);
 
     }
 
@@ -56,15 +71,20 @@ class MedicineController extends Controller
 
     public function destroy($medicine_id)
     {
-        $medicine = Medicine::find($medicine_id);
-        if (!$medicine) {
-            return $this->apiResponse($medicine, 'the post not found',404);
+        if(auth()->user()->role){
+            $medicine = Medicine::find($medicine_id);
+            if (!$medicine) {
+                return $this->apiResponse($medicine, 'the medicine not found',404);
+            }
+            $medicine->delete($medicine_id);
+            if ($medicine) {
+                return $this->apiResponse(null, 'the medicine deleted',200);
+            }
         }
-        $medicine->delete($medicine_id);
-        if ($medicine) {
-            return $this->apiResponse(null, 'the post deleted',200);
-        }
+
+        return $this->apiResponse(null,'you aren\'t admin',403);
     }
+
     public function search($name){
         $medicine=Medicine::where('scientific_name','like','%'.$name.'%')->
         where('trade_name','like','%'.$name.'%')->
@@ -78,20 +98,25 @@ class MedicineController extends Controller
         $request->validate([
             'scientific_name'=>'required|string|max:25',
             'category_id'=>'required',
-            'trade_name'=>'required|string|max:25',
+            'image'=>'required|image',
+            'trade_name'=>'required|string|max:25|unique:medicines',
             'company'=>'required|string|max:25',
             'price'=>'required|max:25']);
 
+        $imagePath = $request->file('image')->store('public/images');
 
+        if(auth()->user()->role) {
+            $medicine = Medicine::find($medicine_id);
+            if (!$medicine) {
+                return $this->apiResponse($medicine, 'the medicine not found', 404);
+            }
+            $medicine->update($request->all());
+            if ($medicine) {
+                return $this->apiResponse($medicine, 'the medicine updated', 201);
+            }
+        }
 
-      $medicine=Medicine::find($medicine_id);
-        if (!$medicine) {
-            return $this->apiResponse($medicine, 'the medicine not found', 404);
-        }
-        $medicine->update($request->all());
-        if ($medicine) {
-            return $this->apiResponse($medicine, 'the medicine updated', 201);
-        }
+        return $this->apiResponse(null,'you aren\'t admin',403);
     }
